+++ packages/primeng/src/index.ts
@@ -5,3 +5,4 @@ export * from './lib/signature-pad.component';
 export * from './lib/time-slots.component';
 export * from './lib/markdown-editor.component';
 export * from './lib/captcha.component';
+export * from './lib/image-crop.component';