+++ packages/primeng/src/lib/ngx-form-engine-primeng.component.ts
@@ -44,6 +44,7 @@ import { SignaturePadComponent } from './signature-pad.component';
 import { TimeSlotsComponent } from './time-slots.component';
 import { MarkdownEditorComponent } from './markdown-editor.component';
 import { CaptchaComponent } from './captcha.component';
+import { ImageCropComponent } from './image-crop.component';
 
 import {
   FieldRegistry,
@@ -97,6 +98,7 @@ import {
     TimeSlotsComponent,
     MarkdownEditorComponent,
     CaptchaComponent,
+    ImageCropComponent,
   ],
 })
 export class NgxJsonFormComponent {